sprate the session usecases in one manger class
