--- src/main/kotlin/logic/usecase/authentication/CleanExpiredSessionUseCase.kt
@@ -1,13 +1,0 @@
-package org.baghdad.logic.usecase.authentication
-
-import org.baghdad.logic.repositories.SessionRepository
-
-class CleanExpiredSessionUseCase(
-    private val sessionRepository: SessionRepository
-) {
-    operator fun invoke() {
-        val session = sessionRepository.loadSession()
-        if (session != null && session.isExpired())
-            sessionRepository.deleteSession()
-    }
-}
